Create Enable-CBT.ps1
Enable CBT on all VMs *and* Templates.

This might be a bit rough around the edges but I think it is better than the dozen or more references and snippets found in the Veeam Forums and KB, as it's at least all in one place. If this has a home here, I can work further on parameters and such.