Custom Cloud Maintenance Mode
PowerShell script that does a couple of things. The first thing is report on the current registry value for the Maintenance Mode and then give you the option to delete the key and return the message to it’s default state. The second thing it does is prompt you enter in the desired custom message and set that in the registry.